Move all 'task' classes to task module
